Move system.advanced to separate plugin
